Extract Typewriter classes
