+++ app/src/main/res/layout/fragment_tutorial_add_quest.xml
@@ -10,7 +10,7 @@
         android:layout_height="0dp"
         android:layout_weight="2.5" />
 
-    <io.ipoli.android.app.ui.TypewriterView
+    <io.ipoli.android.app.ui.typewriter.TypewriterView
         android:id="@+id/tutorial_text"
         style="@style/TextAppearance.AppCompat.Title"
         android:layout_width="match_parent"
